requirements.txt
allure-pytest==2.8.36
allure-python-commons==2.8.36
apipkg==1.5
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
execnet==1.8.0
Faker==8.4.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
